pyproject.toml
[tool.poetry]
name = "photobot"
version = "0.1.0"
description = ""
authors = ["GL1KK <danya_serov05@internet.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
aiogram = "^3.20.0.post0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
